projects/linear_systems_classification/requirements.txt

numpy
scipy
matplotlib
pandas
pyarrow
pyyaml
